сделал модуляцию
