+++ js/data.js
@@ -1,0 +1,31 @@
+const PICTURE_COUNT = 25;
+const AVATAR_COUNT = 6;
+const LIKE_MIN_COUNT = 15;
+const LIKE_MAX_COUNT = 200;
+const COMMENT_COUNT = 25;
+const DESCRIPTIONS = [
+  'Свистать всех на верх',
+  'Тише едешь - дальше будешь.',
+  'Чем дальше в лес, тем больше дров',
+  'Бросить тень на плетень',
+  'Рад бы в рай да грехи не пускают',
+  'Запретный плод сладок',
+];
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом все не плохо. Но  не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула  фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const NAMES = [
+  'Василий',
+  'Андрей',
+  'Макс',
+  'Петя',
+  'Антошик',
+  'Боря',
+];
+
+export { PICTURE_COUNT,AVATAR_COUNT,LIKE_MIN_COUNT,LIKE_MAX_COUNT,COMMENT_COUNT,COMMENT_MESSAGES,DESCRIPTIONS,NAMES}
