+++ js/get-random-integer.js
@@ -1,0 +1,14 @@
+//функция генерирует случайное число
+const getRandomInteger = (a, b) => {
+    const lower = Math.ceil(Math.min(a, b));
+    const upper = Math.floor(Math.max(a, b));
+    const result = Math.random() * (upper - lower + 1) + lower;
+    return Math.floor(result);
+  }
+
+  // функция возвращает случайное число
+const getRandomArrayElement = (array) =>
+array[getRandomInteger(0, array.length - 1)];
+
+
+export { getRandomArrayElement, getRandomInteger };
